+++ src/main/kotlin/view/InputView.kt
@@ -2,10 +2,10 @@ package view
 
 object InputView {
 
-    fun readPurchaseAmount(): Int {
+    fun readPurchaseAmount(): Long {
         return try {
             println("구입금액을 입력해 주세요.")
-            readln().toInt()
+            readln().toLong()
         } catch (e: NumberFormatException) {
             throw IllegalArgumentException("입력값은 정수여야 합니다.")
         }
